+++ packages-next/auth/src/pages/SigninPage.tsx
@@ -1,15 +1,15 @@
 /* @jsx jsx */
 
-import { useState, FormEvent, useRef, useEffect } from 'react';
+import { useState, Fragment, FormEvent, useRef, useEffect } from 'react';
 
-import { jsx, H1, Stack } from '@keystone-ui/core';
+import { jsx, H1, Stack, VisuallyHidden } from '@keystone-ui/core';
 import { Button } from '@keystone-ui/button';
 import { TextInput } from '@keystone-ui/fields';
 import { Notice } from '@keystone-ui/notice';
 
-import { useMutation, gql } from '@keystone-next/admin-ui/apollo';
-import { useRawKeystone, useReinitContext } from '@keystone-next/admin-ui/context';
-import { useRouter } from '@keystone-next/admin-ui/router';
+import { useMutation, gql } from '@keystone-next/keystone/admin-ui/apollo';
+import { useRawKeystone, useReinitContext } from '@keystone-next/keystone/admin-ui/context';
+import { useRouter } from '@keystone-next/keystone/admin-ui/router';
 import { SigninContainer } from '../components/SigninContainer';
 
 type SigninPageProps = {
@@ -101,21 +101,31 @@ export const SigninPage = ({
           </Notice>
         )}
         <Stack gap="medium">
+          <VisuallyHidden as="label" htmlFor="identity">
+            {identityField}
+          </VisuallyHidden>
           <TextInput
+            id="identity"
             name="identity"
             value={state.identity}
             onChange={e => setState({ ...state, identity: e.target.value })}
-            placeholder="Email Address"
+            placeholder={identityField}
             ref={identityFieldRef}
           />
           {mode === 'signin' && (
-            <TextInput
-              name="password"
-              value={state.secret}
-              onChange={e => setState({ ...state, secret: e.target.value })}
-              placeholder="password"
-              type="password"
-            />
+            <Fragment>
+              <VisuallyHidden as="label" htmlFor="password">
+                {secretField}
+              </VisuallyHidden>
+              <TextInput
+                id="password"
+                name="password"
+                value={state.secret}
+                onChange={e => setState({ ...state, secret: e.target.value })}
+                placeholder={secretField}
+                type="password"
+              />
+            </Fragment>
           )}
         </Stack>