+++ packages-next/fields/src/types/timestamp/views/index.tsx
@@ -1,7 +1,7 @@
 /* @jsx jsx */
 import { useState } from 'react';
 
-import { CellContainer, CellLink } from '@keystone-next/admin-ui/components';
+import { CellContainer, CellLink } from '@keystone-next/keystone/admin-ui/components';
 import {
   CardValueComponent,
   CellComponent,