+++ packages-next/types/src/config/index.ts
@@ -1,11 +1,9 @@
-import { IncomingMessage } from 'http';
 import { CorsOptions } from 'cors';
 import type { GraphQLSchema } from 'graphql';
 import type { Config } from 'apollo-server-express';
 
 import type { ImageMode, FileMode, KeystoneContext } from '..';
 
-import { CreateContext } from '../core';
 import type { BaseKeystone } from '../base';
 import { SessionStrategy } from '../session';
 import type { MaybePromise } from '../utils';
@@ -25,7 +23,7 @@ export type KeystoneConfig = {
   db: DatabaseConfig;
   ui?: AdminUIConfig;
   server?: ServerConfig;
-  session?: () => SessionStrategy<any>;
+  session?: SessionStrategy<any>;
   graphql?: GraphQLConfig;
   extendGraphqlSchema?: ExtendGraphqlSchema;
   files?: FilesConfig;
@@ -104,10 +102,8 @@ export type AdminUIConfig = {
   // path?: string;
   getAdditionalFiles?: ((config: KeystoneConfig) => MaybePromise<AdminFileToWrite[]>)[];
   pageMiddleware?: (args: {
-    req: IncomingMessage;
-    session: any;
+    context: KeystoneContext;
     isValidSession: boolean;
-    createContext: CreateContext;
   }) => MaybePromise<{ kind: 'redirect'; to: string } | void>;
 };